Refactor engine.rb to work under new namespace
