+++ lib/defra_ruby.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "defra_ruby/validators"
+require_relative "defra_ruby/validators/engine"
 
 module DefraRuby
   # The Defra Ruby packages namespace
